refactor(core): standardize query key aliases
