+++ packages/core/src/auth/invalidate.ts
@@ -1,18 +1,18 @@
 import type { QueryClient } from '@tanstack/query-core'
-import { createQueryKey as createCanAccessQueryKey } from '../authz/can'
-import { createQueryKey as createPermissionsQueryKey } from '../authz/permissions'
-import { createQueryKey as createCheckQueryKey } from './check'
-import { createQueryKey as createIdentityQueryKey } from './identity'
+import { createQueryKey as genCanAccessQueryKey } from '../authz/can'
+import { createQueryKey as genPermissionsQueryKey } from '../authz/permissions'
+import { createQueryKey as genCheckQueryKey } from './check'
+import { createQueryKey as genIdentityQueryKey } from './identity'
 
 export async function triggerInvalidateAll(
 	queryClient: QueryClient,
 ): Promise<void> {
 	await Promise.all(
 		[
-			createCheckQueryKey(),
-			createIdentityQueryKey(),
-			createPermissionsQueryKey(),
-			createCanAccessQueryKey({}),
+			genCheckQueryKey(),
+			genIdentityQueryKey(),
+			genPermissionsQueryKey(),
+			genCanAccessQueryKey({}),
 		].map(key => queryClient.invalidateQueries(
 			{
 				queryKey: key,
